+++ player/js/elements/helpers/shapes/SVGFillStyleData.js
@@ -1,5 +1,9 @@
 function SVGFillStyleData(elem, data, styleOb){
-	this.o = PropertyFactory.getProp(elem,data.o,0,0.01,elem);
-	this.c = PropertyFactory.getProp(elem,data.c,1,255,elem);
+	this.initDynamicPropertyContainer(elem);
+	this.getValue = this.iterateDynamicProperties;
+	this.o = PropertyFactory.getProp(elem,data.o,0,0.01,this);
+	this.c = PropertyFactory.getProp(elem,data.c,1,255,this);
 	this.style = styleOb;
 }
+
+extendPrototype([DynamicPropertyContainer], SVGFillStyleData);