+++ player/js/elements/helpers/shapes/SVGGradientStrokeStyleData.js
@@ -1,9 +1,10 @@
 function SVGGradientStrokeStyleData(elem, data, styleOb){
-	this.w = PropertyFactory.getProp(elem,data.w,0,null,elem);
-	this.d = new DashProperty(elem,data.d||{},'svg',elem);
+	this.initDynamicPropertyContainer(elem);
+	this.getValue = this.iterateDynamicProperties;
+	this.w = PropertyFactory.getProp(elem,data.w,0,null,this);
+	this.d = new DashProperty(elem,data.d||{},'svg',this);
     this.initGradientData(elem, data, styleOb);
+    this._isAnimated = !!this._isAnimated;
 }
 
-SVGGradientStrokeStyleData.prototype.initGradientData = SVGGradientFillStyleData.prototype.initGradientData;
-SVGGradientStrokeStyleData.prototype.setGradientData = SVGGradientFillStyleData.prototype.setGradientData;
-SVGGradientStrokeStyleData.prototype.setGradientOpacity = SVGGradientFillStyleData.prototype.setGradientOpacity;
+extendPrototype([SVGGradientFillStyleData, DynamicPropertyContainer], SVGGradientStrokeStyleData);