+++ player/js/elements/htmlElements/HShapeElement.js
@@ -11,6 +11,8 @@ function HShapeElement(data,globalData,comp){
     this.itemsData = [];
     //List of items in previous shape tree
     this.processedElements = [];
+    // List of animated components
+    this.animatedContents = [];
     this.shapesContainer = createNS('g');
     this.initElement(data,globalData,comp);
     //Moving any property that doesn't get too much access after initialization because of v8 way of handling more than 10 properties.
@@ -28,6 +30,7 @@ HShapeElement.prototype._renderShapeFrame = HShapeElement.prototype.renderInnerC
 
 HShapeElement.prototype.createContent = function(){
     var cont;
+    this.baseElement.style.fontSize = 0;
     if (this.data.hasMask) {
         this.layerElement.appendChild(this.shapesContainer);
         cont = this.svgElement;
@@ -41,6 +44,7 @@ HShapeElement.prototype.createContent = function(){
     }
 
     this.searchShapes(this.shapesData,this.itemsData,this.prevViewData,this.shapesContainer,0, [], true);
+    this.filterUniqueShapes();
     this.shapeCont = cont;
 };
 
@@ -166,6 +170,13 @@ HShapeElement.prototype.calculateBoundingBox = function(itemsData, boundingBox) 
     }
 }
 
+HShapeElement.prototype.currentBoxContains = function(box) {
+    return this.currentBBox.x <= box.x 
+    && this.currentBBox.y <= box.y 
+    && this.currentBBox.width + this.currentBBox.x >= box.x + box.width
+    && this.currentBBox.height + this.currentBBox.y >= box.y + box.height
+}
+
 HShapeElement.prototype.renderInnerContent = function() {
     this._renderShapeFrame();
 
@@ -180,6 +191,9 @@ HShapeElement.prototype.renderInnerContent = function() {
         tempBoundingBox.width = tempBoundingBox.xMax < tempBoundingBox.x ? 0 : tempBoundingBox.xMax - tempBoundingBox.x;
         tempBoundingBox.height = tempBoundingBox.yMax < tempBoundingBox.y ? 0 : tempBoundingBox.yMax - tempBoundingBox.y;
         //var tempBoundingBox = this.shapeCont.getBBox();
+        if(this.currentBoxContains(tempBoundingBox)) {
+            return;
+        }
         var changed = false;
         if(this.currentBBox.w !== tempBoundingBox.width){
             this.currentBBox.w = tempBoundingBox.width;