+++ player/js/utils/shapes/DashProperty.js
@@ -1,43 +1,34 @@
 function DashProperty(elem, data, renderer) {
     this.elem = elem;
-    this.container = elem;
     this.frameId = -1;
     this.dataProps = createSizedArray(data.length);
     this.renderer = renderer;
-    this._mdf = false;
     this.k = false;
-    this.dynamicProperties = [];
     this.dashStr = '';
     this.dashArray = createTypedArray('float32',  data.length ? data.length - 1 : 0);
     this.dashoffset = createTypedArray('float32',  1);
+    this.initDynamicPropertyContainer(elem);
     var i, len = data.length || 0, prop;
     for(i = 0; i < len; i += 1) {
         prop = PropertyFactory.getProp(elem,data[i].v,0, 0, this);
-        this.k = prop.k ? true : this.k;
+        this.k = prop.k || this.k;
         this.dataProps[i] = {n:data[i].n,p:prop};
     }
     if(!this.k){
         this.getValue(true);
     }
+    this._isAnimated = this.k;
 }
-
-DashProperty.prototype.addDynamicProperty = addDynamicProperty;
 
 DashProperty.prototype.getValue = function(forceRender) {
     if(this.elem.globalData.frameId === this.frameId && !forceRender){
         return;
     }
-    var i = 0, len = this.dataProps.length;
-    this._mdf = forceRender;
     this.frameId = this.elem.globalData.frameId;
-    while(i<len){
-        this.dataProps[i].p.getValue();
-        if(this.dataProps[i].p._mdf){
-            this._mdf = true;
-        }
-        i+=1;
-    }
+    this.iterateDynamicProperties();
+    this._mdf = this._mdf || forceRender;
     if (this._mdf) {
+        var i = 0, len = this.dataProps.length;
         if(this.renderer === 'svg') {
             this.dashStr = '';
         }
@@ -54,3 +45,4 @@ DashProperty.prototype.getValue = function(forceRender) {
         }
     }
 };
+extendPrototype([DynamicPropertyContainer], DashProperty);