+++ player/js/utils/shapes/TrimModifier.js
@@ -9,6 +9,7 @@ TrimModifier.prototype.initModifierProperties = function(elem, data) {
     this.eValue = 0;
     this.getValue = this.processKeys;
     this.m = data.m;
+    this._isAnimated = !!this.s.effectsSequence.length || !!this.e.effectsSequence.length || !!this.o.effectsSequence.length;
 };
 
 TrimModifier.prototype.addShapeToModifier = function(shapeData){
@@ -90,6 +91,8 @@ TrimModifier.prototype.processShapes = function(_isFirstFrame) {
             s = e;
             e = _s;
         }
+        s = Math.round(s*1000)/1000;
+        e = Math.round(e*1000)/1000;
         this.sValue = s;
         this.eValue = e;
     } else {