+++ player/js/utils/text/TextSelectorProperty.js
@@ -4,15 +4,14 @@ var TextSelectorProp = (function(){
     var floor = Math.floor;
 
     function TextSelectorProp(elem,data){
-        this._mdf = false;
+        this._currentTextLength = -1;
         this.k = false;
         this.data = data;
-        this.dynamicProperties = [];
         this.elem = elem;
-        this.container = elem;
         this.comp = elem.comp;
         this.finalS = 0;
         this.finalE = 0;
+        this.initDynamicPropertyContainer(elem);
         this.s = PropertyFactory.getProp(elem,data.s || {k:0},0,0,this);
         if('e' in data){
             this.e = PropertyFactory.getProp(elem,data.e,0,0,this);
@@ -29,8 +28,10 @@ var TextSelectorProp = (function(){
     }
 
     TextSelectorProp.prototype = {
-        addDynamicProperty: addDynamicProperty,
         getMult: function(ind) {
+            if(this._currentTextLength !== this.elem.textProperty.currentData.l.length) {
+                this.getValue();
+            }
             //var easer = bez.getEasingCurve(this.ne.v/100,0,1-this.xe.v/100,1);
             var easer = BezierFactory.getBezierEasing(this.ne.v/100,0,1-this.xe.v/100,1).get;
             var mult = 0;
@@ -102,21 +103,13 @@ var TextSelectorProp = (function(){
             return mult*this.a.v;
         },
         getValue: function(newCharsFlag) {
-            this._mdf = newCharsFlag || false;
-            if(this.dynamicProperties.length){
-                var i, len = this.dynamicProperties.length;
-                for(i=0;i<len;i+=1){
-                    this.dynamicProperties[i].getValue();
-                    if(this.dynamicProperties[i]._mdf){
-                        this._mdf = true;
-                    }
-                }
-            }
-            var totalChars = this.data.totalChars || this.elem.textProperty.currentData.l.length || 0;
+            this.iterateDynamicProperties();
+            this._mdf = newCharsFlag || this._mdf;
+            this._currentTextLength = this.elem.textProperty.currentData.l.length || 0;
             if(newCharsFlag && this.data.r === 2) {
-                this.e.v = totalChars;
+                this.e.v = this._currentTextLength;
             }
-            var divisor = this.data.r === 2 ? 1 : 100 / totalChars;
+            var divisor = this.data.r === 2 ? 1 : 100 / this._currentTextLength;
             var o = this.o.v/divisor;
             var s = this.s.v/divisor + o;
             var e = (this.e.v/divisor) + o;
@@ -129,6 +122,7 @@ var TextSelectorProp = (function(){
             this.finalE = e;
         }
     }
+    extendPrototype([DynamicPropertyContainer], TextSelectorProp);
 
     function getTextSelectorProp(elem, data,arr) {
         return new TextSelectorProp(elem, data, arr);