+++ History.md
@@ -1,3 +1,21 @@
+## V 5.1.11
+- PERFORMANCE: significant improvement on the svg and canvas renderers
+- FIX: floating points fix
+
+## V 5.1.10
+- TEXT LAYERS: font measuring and some text fixes
+- FIX: 2d and 3d layers stack fix
+- FEATURE: added support for assetsPath when using animationData (@kwilliams-curago)
+- FEATURE: added getDuration method. In frames and in seconds.
+- FIX: fixed breaking change from AE 15.1 with text layers
+
+## V 5.1.9
+- EXPRESSIONS: adding name property to shape property group
+- FIX: adding timeout before checking loaded fonts
+- IMPROVEMENT: html elements now hide their base container when off render time bounds
+- FIX: undeclared variable in reverse method
+- FIX: text font issues
+
 ## V 5.1.8
 - FIX: centered tracking
 - FIX: augmenting linearity threshold